refactor: add separate package for simple/paging adapter viewholder
