+++ app/src/main/java/fr/abennsir/poc/album/app/gallery/simple/viewholder/CarouselViewHolder.kt
@@ -1,4 +1,4 @@
-package fr.abennsir.poc.album.app.gallery.viewholder
+package fr.abennsir.poc.album.app.gallery.simple.viewholder
 
 
 import androidx.lifecycle.LifecycleOwner
@@ -6,9 +6,10 @@ import androidx.lifecycle.observe
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewpager2.widget.ViewPager2
-import fr.abennsir.poc.album.app.gallery.adapter.PhotoGalleryDataAdapter
 import fr.abennsir.poc.album.app.gallery.data.NavigationMode
 import fr.abennsir.poc.album.app.gallery.data.UiModel
+import fr.abennsir.poc.album.app.gallery.simple.adapter.PhotoGalleryDataAdapter
+import fr.abennsir.poc.album.app.gallery.viewholder.BaseCarouselViewHolder
 import fr.abennsir.poc.album.app.gallery.viewmodel.PhotoViewModel
 import kotlinx.coroutines.ExperimentalCoroutinesApi
