+++ app/src/main/java/fr/abennsir/poc/album/app/gallery/ui/AlbumFragment.kt
@@ -13,6 +13,7 @@ import fr.abennsir.poc.album.app.core.utils.autoCleared
 import fr.abennsir.poc.album.app.gallery.data.NavigationMode
 import fr.abennsir.poc.album.app.gallery.data.UiModel
 import fr.abennsir.poc.album.app.gallery.injection.ApplicationDependenciesResolver
+import fr.abennsir.poc.album.app.gallery.simple.viewholder.AlbumScreenViewHolder
 import fr.abennsir.poc.album.app.gallery.viewholder.BaseAlbumScreenViewHolder
 import fr.abennsir.poc.album.app.gallery.viewmodel.PhotoViewModel
 import fr.abennsir.poc.album.app.gallery.viewmodel.ViewModelFactory
